Added markdown rendering
This is currently done client side. Ideally the markdown should be
rendered during build time, however that would require us to eject and
manually configure webpack which I would rather not do.

Although not ideal, rendering markdown in the client is fine and the
performance hit is negligible. We also fetch the markdown instead of
baking it into the entrypoint as a string thereby adding an extra
network request which is not ideal.